refactor where read only array lives
