+++ src/storage/mod.rs
@@ -1,8 +1,6 @@
 mod r#async;
-mod read_only;
 mod sync;
 
-pub use r#async::PyAsyncStorage;
-pub(crate) use read_only::{AsyncReadOnly, ReadOnly};
+pub use r#async::{AsyncReadOnly, PyAsyncStorage};
 pub(crate) use sync::PySyncStorage;
-pub use sync::{PyFilesystemStore, PyMemoryStore};
+pub use sync::{PyFilesystemStore, PyMemoryStore, ReadOnly};
